+++ Reminder/app/src/main/java/com/reminder/app/RESTClient.java
@@ -1,36 +1,28 @@
 package com.reminder.app;
 
-/**
- * Created by jordanvega on 5/24/14.
- */
-
 import android.content.Context;
-
 import com.loopj.android.http.*;
-
 import org.apache.http.entity.StringEntity;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.json.JSONArray;
-
-
 import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Arrays;
 import android.util.Log;
 
-
-
 public class RESTClient {
     private static final String REMINDER_URL = "https://flash-energy-585.appspot.com/_ah/api/reminders/v1/reminders/";
+    private static final String CREATE_REMINDER_URL = "createreminder";
+    private static final String LIST_REMINDER_URL = "list";
 
     private static AsyncHttpClient client = new AsyncHttpClient();
+
 
     public static void newReminder(Context context, String username, String title, String[] reminderList, double latitude, double longitude, int urgency) {
         JSONObject jsonParams = new JSONObject();
         JSONArray list = new JSONArray(Arrays.asList(reminderList));
-
         StringEntity entity;
         try {
             jsonParams.put("username", username);
@@ -39,37 +31,28 @@ public class RESTClient {
             jsonParams.put("longitude", longitude);
             jsonParams.put("reminder",list);
             jsonParams.put("urgency", urgency);
-        }
-        catch (JSONException e)
-        {
-            System.out.print(e.toString());
-        }
-        try{
             entity = new StringEntity(jsonParams.toString());
-
-            client.post(context, REMINDER_URL+"createreminder" ,entity, "application/json", new JsonHttpResponseHandler() { });
+            client.post(context, REMINDER_URL+ CREATE_REMINDER_URL, entity, "application/json", new JsonHttpResponseHandler() { });
         }
-        catch (UnsupportedEncodingException e)
-        {
+        catch (Exception e) {
             System.out.print(e.toString());
         }
 
     }
-    public static ArrayList listReminders(Context context, String user){
+    public static ArrayList listReminders(Context context, String user) {
         RequestParams params = new RequestParams();
         params.put("username",user);
-        client.get(context, REMINDER_URL + "list", params, new JsonHttpResponseHandler()  {
-                @Override
-                public void  onSuccess(JSONObject response){
-                        Log.i("testing", response.toString());
-                        //urgency
-                }
-                @Override
-                public void onFailure ( Throwable e, JSONObject errorResponse ) {
-                    String msg = "Object *" + e.toString() + "*" + errorResponse.toString();
-                    Log.i("testing","onFailure: " + msg);
-                }
-            });
+        client.get(context, REMINDER_URL + LIST_REMINDER_URL, params, new JsonHttpResponseHandler() {
+            @Override
+            public void  onSuccess(JSONObject response){
+                Log.i("testing", response.toString());
+              }
+            @Override
+            public void onFailure ( Throwable e, JSONObject errorResponse ) {
+                String msg = "Object *" + e.toString() + "*" + errorResponse.toString();
+                Log.i("testing","onFailure: " + msg);
+            }
+        });
         return null;
     }
 }
