+++ Reminder/app/src/main/res/layout/activity_main.xml
@@ -8,13 +8,6 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context="com.reminderapp.app.MainActivity">
 
-    <com.google.android.gms.common.SignInButton
-        android:id="@+id/sign_in_button"
-        android:layout_width="200dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true" />
-
     <Button
         android:layout_width="200dp"
         android:layout_height="wrap_content"
@@ -29,11 +22,10 @@
         android:layout_width="200dp"
         android:layout_height="wrap_content"
         android:id="@+id/ReminderTitle"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="48dp"
         android:hint="Reminder Title"
-        android:layout_below="@+id/sign_in_button"
-        android:layout_alignLeft="@+id/sign_in_button"
-        android:layout_alignStart="@+id/sign_in_button" />
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true" />
 
     <EditText
         android:layout_width="200dp"
